refactor: modularize Rust/C# codebase, fix warnings, unify CLI contracts
- Fixed cargo incremental-compilation warnings (.cargo/config.toml)
- Split scan-engine/src/scanner.rs into scanner_sync.rs + scanner_progress.rs
- Fixed top_directories empty for shallow scans (gpu-compute/src/scan.rs)
- Eliminated remaining clippy warnings (ollama/helpers.rs, cli/bloat.rs, cli/semantic.rs)
- Converted large C# ViewModels to partial classes (Cleanup, Duplicates, System)
- Split src/cli/mod.rs into subcommands (scan_command, history_command, db_command, settings_command, ai_command)
- Restructured modules into directories: database/scans, app_inventory, origin_tracer, ollama, workflows
- Removed dead scan --channel flag; unified recommendation/priority handling
- Added offline-analysis transparency banner with correct fallback detection
- Updated docs/changelog and INDEX; verified cargo test --workspace --features gpu and WinUI MSBuild pass

+++ src/cli/types.rs
@@ -40,7 +40,7 @@ pub enum StreamEvent {
         total_size_bytes: u64,
         total_size_mb: f64,
         duration_secs: f64,
-        file_types: std::collections::HashMap<String, usize>,
+        file_types: std::collections::HashMap<String, u64>,
         extension_sizes: std::collections::HashMap<String, u64>,
         largest_files: Vec<gui_common::LargestFileEntry>,
         errors: Vec<String>,
